Avoid XML namespaces via private function abuse
The default serialization of ElementTree.tostring() adds `ns0:` type namespace
prefixes to inline svg elements, which doesn't work for HTML.  From
https://developer.mozilla.org/en-US/docs/Web/SVG/Namespaces_Crash_Course

> XML namespaces are not valid in HTML.

> With SVG, [the implied namespace] is http://www.w3.org/2000/svg

So, this commit bypasses the public interface for ElementTree to allow "cleanup"
of namespace declarations and prefixes.  There has to be a better way, right?

Put another way:

`ElementTree.tostring()` turns `<svg>` int `<ns0:svg>`, which breaks rendering.
This is a dirty hack to stop that from happening.